Refactor visualization script into modules
